Compat' with terracotta-management 2.4.0 (moved model classes into a specific model package)
