+++ management/build.gradle
@@ -18,5 +18,5 @@ apply plugin: EhDeploy
 
 dependencies {
   compile project(':api'), project(':core'), project(':impl')
-  compile group: 'org.terracotta', name: 'management-registry', version: parent.managementVersion
+  compile group: 'org.terracotta.management', name: 'management-registry', version: parent.managementVersion
 }
