+++ management/src/main/java/org/ehcache/management/providers/statistics/EhcacheStatistics.java
@@ -27,21 +27,21 @@ import org.terracotta.context.extended.StatisticsRegistry;
 import org.terracotta.context.query.Matcher;
 import org.terracotta.context.query.Matchers;
 import org.terracotta.context.query.Query;
-import org.terracotta.management.capabilities.descriptors.Descriptor;
-import org.terracotta.management.capabilities.descriptors.StatisticDescriptor;
-import org.terracotta.management.capabilities.descriptors.StatisticDescriptorCategory;
-import org.terracotta.management.context.Context;
+import org.terracotta.management.model.capabilities.descriptors.Descriptor;
+import org.terracotta.management.model.capabilities.descriptors.StatisticDescriptor;
+import org.terracotta.management.model.capabilities.descriptors.StatisticDescriptorCategory;
+import org.terracotta.management.model.context.Context;
 import org.terracotta.management.registry.action.ExposedObject;
-import org.terracotta.management.stats.NumberUnit;
-import org.terracotta.management.stats.Sample;
-import org.terracotta.management.stats.Statistic;
-import org.terracotta.management.stats.StatisticType;
-import org.terracotta.management.stats.history.AverageHistory;
-import org.terracotta.management.stats.history.CounterHistory;
-import org.terracotta.management.stats.history.DurationHistory;
-import org.terracotta.management.stats.history.RateHistory;
-import org.terracotta.management.stats.history.RatioHistory;
-import org.terracotta.management.stats.primitive.Counter;
+import org.terracotta.management.model.stats.NumberUnit;
+import org.terracotta.management.model.stats.Sample;
+import org.terracotta.management.model.stats.Statistic;
+import org.terracotta.management.model.stats.StatisticType;
+import org.terracotta.management.model.stats.history.AverageHistory;
+import org.terracotta.management.model.stats.history.CounterHistory;
+import org.terracotta.management.model.stats.history.DurationHistory;
+import org.terracotta.management.model.stats.history.RateHistory;
+import org.terracotta.management.model.stats.history.RatioHistory;
+import org.terracotta.management.model.stats.primitive.Counter;
 import org.terracotta.statistics.OperationStatistic;
 import org.terracotta.statistics.archive.Timestamped;
 import org.terracotta.statistics.extended.Result;
