+++ management/src/main/java/org/ehcache/management/registry/DefaultCollectorService.java
@@ -33,13 +33,13 @@ import org.ehcache.spi.service.Service;
 import org.ehcache.spi.service.ServiceDependencies;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.terracotta.management.context.Context;
-import org.terracotta.management.context.ContextContainer;
-import org.terracotta.management.message.DefaultMessage;
-import org.terracotta.management.notification.ContextualNotification;
+import org.terracotta.management.model.context.Context;
+import org.terracotta.management.model.context.ContextContainer;
+import org.terracotta.management.model.message.DefaultMessage;
+import org.terracotta.management.model.notification.ContextualNotification;
 import org.terracotta.management.registry.MessageConsumer;
 import org.terracotta.management.registry.StatisticQuery;
-import org.terracotta.management.stats.ContextualStatistics;
+import org.terracotta.management.model.stats.ContextualStatistics;
 
 import java.util.ArrayList;
 import java.util.Collection;
