+++ management/src/test/java/org/ehcache/management/providers/actions/EhcacheActionProviderTest.java
@@ -20,11 +20,11 @@ import org.ehcache.config.CacheRuntimeConfiguration;
 import org.ehcache.management.providers.CacheBinding;
 import org.hamcrest.Matcher;
 import org.junit.Test;
-import org.terracotta.management.call.Parameter;
-import org.terracotta.management.capabilities.context.CapabilityContext;
-import org.terracotta.management.capabilities.descriptors.CallDescriptor;
-import org.terracotta.management.capabilities.descriptors.Descriptor;
-import org.terracotta.management.context.Context;
+import org.terracotta.management.model.call.Parameter;
+import org.terracotta.management.model.capabilities.context.CapabilityContext;
+import org.terracotta.management.model.capabilities.descriptors.CallDescriptor;
+import org.terracotta.management.model.capabilities.descriptors.Descriptor;
+import org.terracotta.management.model.context.Context;
 
 import java.util.Arrays;
 import java.util.Collection;
