+++ management/src/test/java/org/ehcache/management/registry/DefaultManagementRegistryServiceTest.java
@@ -25,13 +25,13 @@ import org.terracotta.management.registry.StatisticQuery;
 import org.ehcache.management.config.EhcacheStatisticsProviderConfiguration;
 import org.hamcrest.Matchers;
 import org.junit.Test;
-import org.terracotta.management.call.ContextualReturn;
-import org.terracotta.management.capabilities.Capability;
-import org.terracotta.management.context.Context;
-import org.terracotta.management.stats.ContextualStatistics;
-import org.terracotta.management.stats.Sample;
-import org.terracotta.management.stats.history.CounterHistory;
-import org.terracotta.management.stats.primitive.Counter;
+import org.terracotta.management.model.call.ContextualReturn;
+import org.terracotta.management.model.capabilities.Capability;
+import org.terracotta.management.model.context.Context;
+import org.terracotta.management.model.stats.ContextualStatistics;
+import org.terracotta.management.model.stats.Sample;
+import org.terracotta.management.model.stats.history.CounterHistory;
+import org.terracotta.management.model.stats.primitive.Counter;
 
 import java.io.Serializable;
 import java.util.ArrayList;
